chore: add types to components/list-players
