+++ src/components/list-players/ListPlayers.styles.ts
